test(pdf): extract quality equation family (#252)
* test(research): extract equation quality tests

* refactor(research): extract visual text quality tests

* refactor(research): extract OCR asset quality tests

* refactor(research): extract visual obligation quality tests

* refactor(research): extract semantic flow quality tests

---------
